bg_image upload and edit for all sections

+++ app/Http/Controllers/Api/v1/HeroController.php
@@ -37,15 +37,22 @@ class HeroController extends Controller
                 Storage::disk('public')->delete($hero->image_url);
             }
 
+            $image = $request->file('image_url');
+            $fileName = time() . '_' . uniqid() . '.' . $image->getClientOriginalExtension();
+
+            $image->storeAs('heroContent', $fileName, 'public');
+            $data['image_url'] = 'heroContent/' . $fileName;
             // Store new image
-            $data['image_url'] = $request->file('image_url')->store('heroContent', 'public');
+            // $data['image_url'] = $fileName->store('heroContent', 'public');
         } 
-        elseif 
-            ($request->filled('image_url') === false && $hero->image_url) {
+        elseif ($request->filled('image_url') === false && $hero->image_url) 
+        {
             if (Storage::disk('public')->exists($hero->image_url)) {
                 Storage::disk('public')->delete($hero->image_url);
             }
             $data['image_url'] = null;
+        } else {
+            unset($data['image_url']);
         }
 
         $hero = $this->heroService->updateHeroContent($id, $data);
